requirements.txt
alembic==1.16.4
altgraph==0.17.4
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
babel==2.17.0
backrefs==5.9
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coverage==7.10.4
Django==5.2.5
django-cleanup==9.0.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
ghp-import==2.1.0
git-filter-repo==2.47.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.17
mkdocs-material-extensions==1.3.1
mysqlclient==2.2.7
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
pymdown-extensions==10.16.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyyaml_env_tag==1.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
setuptools==80.9.0
shiboken6==6.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
